Pessoa to Contatos link and adjustments
TODO : Implement screens inserting data. The screens for deleting data is going to extend from the current "consulta" screens, and screens for updating data are going to use the same ones as for inserting data, except the current information about the person/contact/contact type are going to be loaded as well.